backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jambox"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "aiosqlite>=0.20",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "authlib>=1.3",
    "itsdangerous>=2.2",
    "python-jose[cryptography]>=3.3",
    "aiofiles>=23.2",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
